% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{check_names}
\alias{check_names}
\title{\code{check_names} to check plant species names in Wisconsin for typos and synonyms using UWSP herbarium as baseline.}
\usage{
check_names(sp.to.check, sp.dataset = sp_syn_uwsp)
}
\arguments{
\item{sp.to.check}{A vector of plant species to check.}

\item{sp.dataset}{species names and their synonyms from UWSP herbarium website.}
}
\value{
\itemize{
  \item sp.origin the original species names.
  \item sp.spell the closest correct species names after check for typos, if any.
  \item sp.final the updated species names after check for synonyms, if any.
  \item spell.correct logical, are the species names spelled correctly?
  \item syc.correct logical, are the species names updated?
}
}
\description{
\code{check_names} to check plant species names in Wisconsin for typos and synonyms using UWSP herbarium as baseline.
}
\examples{
check_names(sp.to.check = c("Acer rubrum", "Quercus abla"))
}
\author{
Daijiang Li
}
