% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{dist_to_df}
\alias{dist_to_df}
\title{\code{dist_to_df} convert a distance class object or a matrix to a data frame.}
\usage{
dist_to_df(x)
}
\arguments{
\item{x}{a object of class "dist" or a matrix.}
}
\value{
a data frame with three columns: site1, site2, distance.
}
\description{
\code{dist_to_df} convert a distance class object or a matrix to a data frame.
}
\author{
Daijiang Li
}
