% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{get_pd_alpha}
\alias{get_pd_alpha}
\title{calculate alpha phylogenetic diversity}
\usage{
get_pd_alpha(samp_wide, tree, samp_long, null.model = TRUE, null.type = 0,
  n.item = 999, abund.weight = TRUE, ...)
}
\arguments{
\item{samp_wide}{wide version of samp_long, row.names are sites, colnames are species}

\item{tree}{a phylogeny with class of 'phylo'}

\item{samp_long}{a 3-column data frame, site, freq, sp}

\item{null.model}{whether to run null models for MPD and MNTD?}

\item{null.type}{if null.model is TRUE, which null model to use?See ?phylocomr::ph_comstruct for details}

\item{n.item}{the number of randomization}

\item{abund.weight}{should abundance information used when calculating pd?}
}
\value{
a data frame
}
\description{
A function to calculate a bunch of phylo diversity: Faith's PD, MPD, VPD (variance of pairwise distance), MNTD, PSV/PSE.
}
