% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{mvpd}
\alias{mvpd}
\title{calculate MPD and VPD (mean and variance of pairwise distance)}
\usage{
mvpd(samp, dis, abundance.weighted = FALSE)
}
\description{
calculate MPD and VPD (mean and variance of pairwise distance)
}
