% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{pd2}
\alias{pd2}
\title{Faith's PD}
\usage{
pd2(comm, tree, include.root = TRUE, comm_long)
}
\arguments{
\item{comm}{a site by species data frame, site names as row names}

\item{tree}{a phylogeny of class "phylo"}

\item{include.root}{whether include root in picante::pd; phylocomr::ph_pd always include root}

\item{comm_long}{long format of comm, 3-columns: site, freq, sp; optional.}
}
\value{
a data frame of PD for each site
}
\description{
Calculate faith's pd, this is a wrapper of phylocomr::ph_pd and picante::pd
}
