% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rand_test}
\alias{rand_test}
\title{\code{rand_test} to perform randomization test between two vectors.}
\usage{
rand_test(x, y, n = 1000)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{n}{the number of randomization, default is 1000.}
}
\value{
a data frame with mean, rank, p.value, etc.
}
\description{
\code{rand_test} to perform randomization test between two vectors.
}
\author{
Daijiang Li
}
