% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rm_site_noobs}
\alias{rm_site_noobs}
\title{\code{rm_site_noobs} remove site that has no obsrevations of any species.}
\usage{
rm_site_noobs(df)
}
\arguments{
\item{df}{a data frame in wide form, i.e. site by species data frame, with site names as row name.}
}
\value{
a data frame.
}
\description{
\code{rm_site_noobs} remove site that has no obsrevations of any species.
}
\author{
Daijiang Li
}
