% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rm_sp_noobs}
\alias{rm_sp_noobs}
\title{\code{rm_sp_noobs} remove species that not observed in any site.}
\usage{
rm_sp_noobs(df)
}
\arguments{
\item{df}{a data frame in wide form, i.e. site by species data frame, with site names as row name.}
}
\value{
a data frame.
}
\description{
\code{rm_sp_noobs} remove species that not observed in any site.
}
\author{
Daijiang Li
}
