% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{simpleCap}
\alias{simpleCap}
\title{\code{simpleCap} to make the first letter to be upper case.}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{a vector of species names.}
}
\value{
a vector.
}
\description{
\code{simpleCap} to make the first letter to be upper case.
}
\author{
Daijiang Li
}
