% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{tnrs_match_names_2}
\alias{tnrs_match_names_2}
\title{match taxa names with the Open Tree of Life}
\usage{
tnrs_match_names_2(taxa, n_per_req = 20, ...)
}
\arguments{
\item{taxa}{a vector of species names to match}

\item{n_per_req}{number of species to match per requery, must be <= 250}
}
\value{
if the length of taxa is smaller than n_per_req, then a data frame will be returned. otherwise, a list
}
\description{
based on rotl::tnrs_match_names(), which only allow <= 250 species each time. This function will try to do batch match
}
