% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{var_to_rownames}
\alias{var_to_rownames}
\title{\code{var_to_rownames} convert a column of a data frame to be row name.}
\usage{
var_to_rownames(df, var = "site")
}
\arguments{
\item{df}{a data frame.}

\item{var}{the column you want to put as row name, the name is quoted.}
}
\value{
a data frame.
}
\description{
\code{var_to_rownames} convert a column of a data frame to be row name.
}
\author{
Daijiang Li
}
