% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{get_pd_alpha}
\alias{get_pd_alpha}
\title{calculate alpha phylogenetic diversity}
\usage{
get_pd_alpha(samp_wide, tree, samp_long, null.model.phylomeasures = TRUE,
  null.model.phylocom = FALSE, null.type.phylomeasures = "uniform",
  null.type.phylocom = 0, null.type.picante = "taxa.labels", n.item = 999,
  abund.weight = FALSE, verbose = TRUE, vpd = FALSE, ...)
}
\arguments{
\item{samp_wide}{wide version of samp_long, row.names are sites, colnames are species}

\item{tree}{a phylogeny with class of 'phylo'}

\item{samp_long}{a 3-column data frame, site, freq, sp}

\item{null.type.phylomeasures}{if null.model is TRUE, which null model to use for PhyloMeasure?}

\item{null.type.phylocom}{if null.model is TRUE, which null model to use for Phylocom? 
- 0: phylogeny shuffle. 
- 1: maintain site richness and draw from species actually observed in sites. 
- 2: maintain site richness and draw from the phylogeny
- 3: independent swap. 
See ?phylocomr::ph_comstruct for details}

\item{null.type.picante}{if null.model is TRUE, which null model to use for Picante? Not used yet.}

\item{n.item}{the number of randomization}

\item{abund.weight}{should abundance information used when calculating pd with Phylocom/Picante? Default is FALSE}

\item{verbose}{do you want to see relevant information?}

\item{vpd}{to calculate vpd (varance of pairwise distance) or not?}

\item{null.model}{whether to run null models for MPD and MNTD?}

\item{null.model.pd.root}{whether to run null models for rooted PD?}
}
\value{
a data frame
}
\description{
A function to calculate a bunch of phylo diversity: Faith's PD, MPD, VPD (variance of pairwise distance), MNTD, PSV/PSE.
Results of MPD/MNTD from PhyloMeasures are equal with those from Phylocom/Picante with abundance.weight = FALSE
Results of PD from PhyloMeasures are unrooted; PD from Phylocom are rooted.
}
