% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{get_pd_beta}
\alias{get_pd_beta}
\title{calculate pairwise beta phylogenetic diversity}
\usage{
get_pd_beta(samp_wide, tree, samp_long, get.unif = TRUE, get.mpd = TRUE,
  get.mntd = TRUE, get.pcd = TRUE, null.model.phylocom = TRUE,
  null.type.phylocom = 0, n.item = 999, abund.weight = FALSE,
  null.model.mpd.phylomeasures = TRUE,
  null.model.mntd.phylomeasures = FALSE, verbose = TRUE,
  verbose.mntd.null = FALSE, ...)
}
\arguments{
\item{samp_wide}{wide version of samp_long, row.names are sites, colnames are species}

\item{tree}{a phylogeny with class of 'phylo'}

\item{samp_long}{a 3-column data frame, site, freq, sp}

\item{get.unif}{whether to calculate pairwise UniFrac}

\item{get.mpd}{whether to calculate pairwise mpd_beta}

\item{get.mntd}{whether to calculate pairwise mntd_beta}

\item{get.pcd}{calculate PCD or not? Can be time consuming.}

\item{null.model.phylocom}{whether to run null models for MPD and MNTD with Phylocom?}

\item{null.type.phylocom}{if null.model.phylocom is TRUE, which null model to use? See ?phylocomr::ph_comstruct for details}

\item{n.item}{the number of randomization with Phylocom}

\item{abund.weight}{should abundance information used when calculating MPD/MNTD with Phylocom? Note: mntd_beta with Phylocom seems not reliable.}

\item{null.model.mpd.phylomeasures}{whether to run null models for MPD with PhyloMeasures?}

\item{null.model.mntd.phylomeasures}{whether to run null models for MNTD with PhyloMeasures?}

\item{verbose}{do you want to see relevant information?}

\item{verbose.mntd.null}{do you want to see relevant information about null models of MNTD with PhyloMeasures?}
}
\value{
a data frame
}
\description{
A function to calculate a bunch of pairwise beta phylo diversity: 
unifraction, MPD, MNTD, PCD
}
