% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{pcd2}
\alias{pcd2}
\title{pairwise site phylogenetic community dissimilarity (PCD) within a community}
\usage{
pcd2(comm, tree, expectation = NULL, cpp = TRUE, unif_dim = 1000,
  verbose = TRUE)
}
\arguments{
\item{comm}{site by species data frame or matrix, sites as rows.}

\item{tree}{a phylogeny for species}

\item{expectation}{nsp_pool, psv_bar, psv_pool, and nsr calculated from \code{pcd_pred()}.}

\item{cpp}{whether to use loops written with c++, default is TRUE}

\item{unif_dim}{the number of cells (nrow * ncol) of the comm, if it is less than unif_dim, then calculate unifrac and phylosor; these functions are very slow.}

\item{verbose}{do you want to see the progress?}
}
\value{
a list of a variety of pairwise dissimilarities.
}
\description{
Calculate pairwise site PCD, users can specify expected values from \code{pcd_pred()}.
}
\examples{
x1 = pcd_pred(comm_old = read.csv('data/li_2015_old.csv', row.names = 1, check.names = F),
              comm_new = read.csv('data/li_2015_new.csv', row.names = 1, check.names = F),
                 tree = ape::read.tree('data/phy.tre'), reps = 100)
pcd2(comm = read.csv('data/li_2015_old.csv', row.names = 1, check.names = F),
      tree = ape::read.tree('data/phy.tre'), 
      expectation = x1)
}
