% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{phylo_betapart}
\alias{phylo_betapart}
\title{Phylogenetic beta diversity partition}
\usage{
phylo_betapart(comm, tree, index.family = "jaccard", pairwise = TRUE)
}
\arguments{
\item{comm}{a site by species data frame, site names as row names}

\item{tree}{a phylogeny of class "phylo"}

\item{index.family}{"jaccard" or "sorensen"}

\item{pairwise}{calculate pairwise beta diversity or multisite beta diversity}
}
\value{
a list of three distance matrix. For jaccard, phylo.beta.jtu is the turnover-fraction of Jaccard, phylo.beta.jne is the nestedness-fraction.
For sorensen, phylo.beta.sim is the turnover part measured as Simpson derived pairwise dissimilarity, phylo.beta.sne is the nestedness-fraction.
}
\description{
Calculate Phylogenetic beta diversity and its partition, adapted from betapart::phylo.belt.xx()
}
