% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wide_by_site_rarefy}
\alias{wide_by_site_rarefy}
\title{\code{wide_by_site_rarefy} to resample quadrats from survey data}
\usage{
wide_by_site_rarefy(df, n = 1000, q = 20)
}
\arguments{
\item{df}{data frame of vegtation data. Should have site, quad, sp columns.
It will call wide_by_site function inside.}

\item{n}{times of ramdom sampling from quadrats of 2000s data for each site.}

\item{q}{how many quadrats to sample out for 2000s data for all sites.}
}
\value{
a list of lists. A list of sites, each sites then has n ramdon rampled site by sp data frames
(formed a list for that site).
}
\description{
\code{wide_by_site_rarefy} to resample quadrats from survey data
}
\author{
Daijiang Li
}

